fix(state-architecture): Refactor commands and controller to use the same package for better protected behaviour
