+++ adapter/src/main/java/fr/delivrooom/adapter/in/javafxgui/controller/Command.java
@@ -1,7 +1,8 @@
-package fr.delivrooom.adapter.in.javafxgui.command;
+package fr.delivrooom.adapter.in.javafxgui.controller;
 
 
 public interface Command {
+
     void execute();
 
     void undo();
