+++ adapter/src/main/java/fr/delivrooom/adapter/in/javafxgui/controller/CommandManager.java
@@ -1,4 +1,4 @@
-package fr.delivrooom.adapter.in.javafxgui.command;
+package fr.delivrooom.adapter.in.javafxgui.controller;
 
 import java.util.Stack;
