+++ adapter/src/main/java/fr/delivrooom/adapter/in/javafxgui/controller/CommandRemoveDelivery.java
@@ -1,9 +1,8 @@
-package fr.delivrooom.adapter.in.javafxgui.command;
+package fr.delivrooom.adapter.in.javafxgui.controller;
 
-import fr.delivrooom.adapter.in.javafxgui.controller.AppController;
 import fr.delivrooom.application.model.Delivery;
 
-public record RemoveDeliveryCommand(AppController controller, Delivery delivery) implements Command {
+public record CommandRemoveDelivery(AppController controller, Delivery delivery) implements Command {
 
     @Override
     public void execute() {
