+++ adapter/src/main/java/fr/delivrooom/adapter/in/javafxgui/controller/StateDeliveriesLoaded.java
@@ -5,16 +5,16 @@ import fr.delivrooom.application.model.Intersection;
 import java.net.URL;
 
 /**
- * Select intersection state - map and deliveries have been loaded, the use now wants to select an intersection.
- * Allows the user to select an intersection on the map canvas.
+ * Deliveries loaded state - both map and deliveries have been loaded.
+ * Allows opening new map or deliveries files.
  */
-public record SelectIntersectionState(AppController controller) implements State {
+public record StateDeliveriesLoaded(AppController controller) implements State {
 
     @Override
     public void openMapFile(URL url) {
         try {
             controller.loadMapFile(url);
-            controller.setState(new MapLoadedState(controller));
+            controller.setState(new StateMapLoaded(controller));
         } catch (Exception e) {
             controller.showError("Error loading map", e.getMessage());
             e.printStackTrace();
@@ -25,7 +25,7 @@ public record SelectIntersectionState(AppController controller) implements State
     public void openDeliveriesFile(URL url) {
         try {
             controller.loadDeliveriesFile(url);
-            controller.setState(new DeliveriesLoadedState(controller));
+            controller.setState(new StateDeliveriesLoaded(controller));
         } catch (Exception e) {
             controller.showError("Error loading deliveries", e.getMessage());
             e.printStackTrace();
@@ -34,22 +34,27 @@ public record SelectIntersectionState(AppController controller) implements State
 
     @Override
     public void selectIntersection(Intersection intersection) {
-        controller.selectIntersection(intersection);
-        controller.setState(new DeliveriesLoadedState(controller));
+        controller.showError("Can’t select intersection", "Unable to select intersection for now.");
     }
 
     @Override
     public void requestIntersectionSelection() {
-        // Nothing to do here, already in this state.
+        controller.setState(new StateSelectIntersection(controller));
     }
 
     @Override
     public String getStateName() {
-        return "SelectIntersectionState";
+        return "DeliveriesLoadedState";
     }
 
     @Override
     public void requestCalculateTour() {
-        controller.showError("Cannot calculate tour", "Please select an intersection first.");
+        try {
+            controller.calculateTour();
+        } catch (Exception e) {
+            controller.showError("Error calculating tour", e.getMessage());
+            e.printStackTrace();
+        }
     }
+
 }
